+++ lib/exzeitable/filter.ex
@@ -37,17 +37,22 @@ defmodule Exzeitable.Filter do
   end
 
   @doc "Returns true if any of the fields have search enabled"
+  @spec search_enabled?(map) :: boolean
   def search_enabled?(%{fields: fields}) do
     fields
     |> Enum.filter(fn {_k, field} -> Map.get(field, :search) end)
     |> Enum.any?()
   end
 
+  @doc "All fields where a key value is true"
+  @spec fields_where([keyword], atom) :: [keyword]
   def fields_where(fields, attribute) do
     fields
     |> Enum.filter(fn {_k, field} -> Map.get(field, attribute) end)
   end
 
+  @doc "All fields where a key value is false"
+  @spec fields_where_not([keyword], atom) :: [keyword]
   def fields_where_not(fields, attribute) do
     fields
     |> Enum.reject(fn {_k, field} -> Map.get(field, attribute) end)
@@ -62,6 +67,7 @@ defmodule Exzeitable.Filter do
   end
 
   # If virtual: true, a number of other options have to be overridden
+  @spec merge_fields(keyword) :: keyword
   defp merge_fields([virtual: true] = field) do
     @default_fields
     |> Keyword.merge(field)