+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.1.2 (2019-10-11)
+
+* Added ability to show and hide 'show buttons' for columns
+* Disabled use of enter key with search box (created a CSRF error)
+* Added specs
+* Added documentation
+
 ## v0.1.1 (2019-10-09)
 
 * Fix iOS touch issues